fixup! [K/JS] Add exportability of ReceiveChannel and SendChannel for Kotlin/JS with conversion by `AsyncIterator` protocol
