+++ kotlinx-coroutines-core/common/src/utils/JsOptionalExport.kt
@@ -1,0 +1,5 @@
+package kotlinx.coroutines.utils
+
+@OptionalExpectation
+@Target(AnnotationTarget.CLASS)
+internal expect annotation class JsOptionalExport(val couldBeConvertedToExplicitExport: Boolean)
