+++ kotlinx-coroutines-core/js/src/utils/JsOptionalExport.kt
@@ -1,0 +1,4 @@
+package kotlinx.coroutines.utils
+
+@Suppress("INVISIBLE_REFERENCE")
+internal actual typealias JsOptionalExport = JsImplicitExport
